+++ Akade.IndexedSet.Tests/Data/TestData.cs
@@ -1,3 +1,5 @@
 ﻿namespace Akade.IndexedSet.Tests.Data;
 
 public record TestData(int PrimaryKey, int IntProperty, Guid GuidProperty, string StringProperty);
+
+public record DenormalizedTestData(int PrimaryKey, IEnumerable<int> IntList);
