+++ Akade.IndexedSet/Akade.IndexedSet.csproj
@@ -6,7 +6,7 @@
 		<Nullable>enable</Nullable>
 		<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
 		<GenerateDocumentationFile>True</GenerateDocumentationFile>
-		<Version>0.1.1</Version>
+		<Version>0.2.0</Version>
 		<VersionSuffix>beta</VersionSuffix>
 		<Description>Provides an In-Memory data structure, the IndexedSet, that allows to easily add indices to allow efficient querying. Currently supports unique and non-unique indices as well as range indices for single attribute, compound or computed keys.</Description>
 		<Copyright>Copyright © Akade 2022</Copyright>
